requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.7.2
requests==2.31.0